requirements.txt
pyyaml~=6.0
torch~=2.0.0
numpy~=1.24.2
torchvision~=0.15.1
matplotlib~=3.7.1
pillow~=9.5.0
opencv-python~=4.7.0.72
